Refactor and fix exception handling for coroutines
